+++ backintime/broker/base.py
@@ -1,173 +1,237 @@
+"""
+Orders only hold required data, but the desired functions 
+must be implemented somewhere else. 
+All fields are public. It is up to a broker implementation to set 
+`status` and provide TP/SL orders features:
+
+    - Multiple TP/SL orders can be posted for the same position, 
+      so the summarised value of opened TP/SL orders can be greater
+      than the opened position. 
+      In this case, only the first TP/SL order whose conditions
+      are satisfied will be executed. Others must be cancelled.
+
+    - When TP/SL order is activated (target_price meets market), 
+      it must be treated as Market/Limit order,
+      depending on whether `order_price` was provided.
+"""
 import typing as t
-from datetime import datetime
-from dataclasses import dataclass
 from abc import ABC, abstractmethod
-from decimal import Decimal    # https://docs.python.org/3/library/decimal.html
-from .orders import (
-    OrderSide,
-    Order, 
-    OrderType,
-    OrderStatus, 
-    OrderFactory,
-    StrategyOrder,
-    MarketOrder, 
-    LimitOrder, 
-    StopLossOrder,
-    TakeProfitOrder,
-    MarketOrderFactory,
-    LimitOrderFactory,
-    StopLossFactory,
-    TakeProfitFactory
-)
+from enum import Enum
+from decimal import Decimal, ROUND_FLOOR   # https://docs.python.org/3/library/decimal.html
+from dataclasses import dataclass
+from datetime import datetime
 
 
-class BrokerException(Exception):
-    """Base class for all broker-related exceptions."""
-    pass
+class OrderSide(Enum):
+    BUY = "BUY"
+    SELL = "SELL"
+
+    def __str__(self) -> str:
+        return self.value
 
 
-class OrderSubmissionError(BrokerException):
-    """Generic exception for order submission error."""
-    pass
+class OrderStatus(Enum):
+    CREATED = "CREATED"
+    CANCELLED = "CANCELLED"
+    EXECUTED = "EXECUTED"
+    # Only for TP/SL orders
+    SYS_CANCELLED = "SYS_CANCELLED"
+    ACTIVATED = "ACTIVATED"
+
+    def __str__(self) -> str:
+        return self.value
 
 
-class InvalidOrderData(OrderSubmissionError):
-    """Order submission failed because order data is invalid."""
-    pass
+class OrderType(Enum):
+    MARKET = "MARKET"
+    LIMIT = "LIMIT"
+    STOP_LOSS = "STOP_LOSS"
+    STOP_LOSS_LIMIT = "STOP_LOSS_LIMIT"
+    TAKE_PROFIT = "TAKE_PROFIT"
+    TAKE_PROFIT_LIMIT = "TAKE_PROFIT_LIMIT"
+
+    def __str__(self) -> str:
+        return self.value
 
 
-class InsufficientFunds(OrderSubmissionError):
-    """Order submission failed due to insufficient funds."""
-    pass
+class OrderOptions:
+    def __init__(self, 
+                 order_type: OrderSide, 
+                 amount: t.Optional[Decimal] = None,
+                 percentage_amount: t.Optional[Decimal] = None,
+                 order_price: t.Optional[Decimal] = None):
+        self.order_type = order_type
+        self.amount = amount
+        self.percentage_amount = percentage_amount
+        self.order_price = order_price
 
 
-class OrderCancellationError(BrokerException):
-    """Generic exception for order cancellation error."""
-    pass
-
-
-# Implementation for orders & trades
-class OrderInfo:
-    """
-    Wrapper around `Order` that provides a read-only view
-    into the wrapped `Order` data.
-    """
-    def __init__(self, order_id: int, order: Order):
-        self._order_id = order_id
-        self._order = order
-
+class OrderInfo(ABC):
     @property
+    @abstractmethod
     def order_id(self) -> int:
-        return self._order_id
+        pass
 
     @property
+    @abstractmethod
     def order_type(self) -> OrderType:
-        return self._order.order_type
+        pass
 
     @property
+    @abstractmethod
     def order_side(self) -> OrderSide:
-        return self._order.side
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def amount(self) -> Decimal:
-        return self._order.amount
+        pass
 
     @property
+    @abstractmethod
     def date_created(self) -> datetime:
-        return self._order.date_created
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def order_price(self) -> t.Optional[Decimal]:
-        return self._order.order_price
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def status(self) -> OrderStatus:
-        return self._order.status
+        pass
 
     @property
+    @abstractmethod
     def date_updated(self) -> datetime:
-        return self._order.date_updated
+        pass
 
     @property
+    @abstractmethod
     def fill_price(self) -> t.Optional[Decimal]:
-        return self._order.fill_price
+        pass
 
     @property
+    @abstractmethod
     def trading_fee(self) -> t.Optional[Decimal]:
-        return self._order.trading_fee
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def is_unfulfilled(self) -> bool:
-        return not self._order.fill_price
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def is_canceled(self) -> bool:
-        return self._order.status is OrderStatus.CANCELLED or \
-                self._order.status is OrderStatus.SYS_CANCELLED
+        pass
 
-    @property 
+    @property
+    @abstractmethod
     def is_executed(self) -> bool:
-        return self._order.status is OrderStatus.EXECUTED
+        pass
+
+
+class MarketOrderOptions(OrderOptions):
+    def __init__(self, 
+                 order_side: OrderSide, 
+                 amount: t.Optional[Decimal] = None,
+                 percentage_amount: t.Optional[Decimal] = None):
+        self.order_side = order_side
+        super().__init__(OrderType.MARKET, amount, percentage_amount)
+
+
+class MarketOrderInfo(OrderInfo): 
+    pass
 
 
 class StrategyOrderInfo(OrderInfo):
     @property
+    @abstractmethod
     def trigger_price(self) -> Decimal:
-        return self._order.trigger_price
+        pass
 
     @property
+    @abstractmethod
     def is_activated(self) -> bool:
-        return self._order.status is OrderStatus.ACTIVATED
+        pass
 
 
-@dataclass
-class StrategyOrders:
-    take_profit_id: t.Optional[int] = None
-    stop_loss_id: t.Optional[int] = None
+class TakeProfitOptions(OrderOptions):
+    def __init__(self, 
+                 trigger_price: Decimal, 
+                 amount: t.Optional[Decimal] = None,
+                 percentage_amount: t.Optional[Decimal] = None,
+                 order_price: t.Optional[Decimal] = None):
+        self.trigger_price = trigger_price
+        order_type = OrderType.TAKE_PROFIT_LIMIT if order_price \
+                        else OrderType.TAKE_PROFIT
+        super().__init__(order_type, amount, percentage_amount, order_price)
+
+
+class TakeProfitInfo(StrategyOrderInfo):
+    pass
+
+
+class StopLossOptions(OrderOptions):
+    def __init__(self, 
+                 trigger_price: Decimal, 
+                 amount: t.Optional[Decimal] = None,
+                 percentage_amount: t.Optional[Decimal] = None,
+                 order_price: t.Optional[Decimal] = None):
+        self.trigger_price = trigger_price
+        order_type = OrderType.STOP_LOSS_LIMIT if order_price \
+                        else OrderType.STOP_LOSS
+        super().__init__(order_type, amount, percentage_amount, order_price)
+
+
+class StopLossInfo(StrategyOrderInfo):
+    pass
+
+
+class LimitOrderOptions(OrderOptions):
+    def __init__(self, 
+                 order_side: OrderSide, 
+                 order_price: Decimal, 
+                 amount: t.Optional[Decimal] = None,
+                 percentage_amount: t.Optional[Decimal] = None,
+                 take_profit: t.Optional[TakeProfitOptions] = None,
+                 stop_loss: t.Optional[StopLossOptions] = None):
+        self.order_side = order_side
+        self.take_profit = take_profit
+        self.stop_loss = stop_loss
+        super().__init__(OrderType.LIMIT, amount, percentage_amount, order_price)
 
 
 class LimitOrderInfo(OrderInfo):
-    def __init__(self, order_id: int, order, strategy_orders: StrategyOrders):
-        self._strategy_orders = strategy_orders
-        super().__init__(order_id, order)
-
-    @property 
-    def take_profit(self) -> t.Optional[StrategyOrderInfo]:
-        take_profit = self._order.take_profit
-        if take_profit:
-            return StrategyOrderInfo(self._strategy_orders.take_profit_id,
-                                     take_profit)
+    @property
+    @abstractmethod
+    def take_profit(self) -> t.Optional[TakeProfitInfo]:
+        pass
 
     @property
-    def stop_loss(self) -> t.Optional[StrategyOrderInfo]:
-        stop_loss = self._order.stop_loss
-        if stop_loss:
-            return StrategyOrderInfo(self._strategy_orders.stop_loss_id,
-                                     stop_loss)
+    @abstractmethod
+    def stop_loss(self) -> t.Optional[StopLossInfo]:
+        pass
 
 
-class Trade:
-    def __init__(self, 
-                 trade_id: int, 
-                 order_info: OrderInfo, 
-                 result_balance: Decimal):
-        self._trade_id = trade_id
-        self._order_info = order_info
-        self._result_balance = result_balance
-
+class TradeInfo(ABC):
     @property
+    @abstractmethod
     def trade_id(self) -> int:
-        return self._trade_id
+        pass
 
     @property
+    @abstractmethod
     def order(self) -> OrderInfo:
-        # Info about the order
-        return self._order_info
-    
+        pass
+
     @property
+    @abstractmethod
     def result_balance(self) -> Decimal:
         # fiat balance at the moment of order execution
-        return self._result_balance
+        pass
 
 
 class AbstractBalance:
@@ -194,6 +258,31 @@ class AbstractBalance:
     def crypto_balance(self) -> Decimal:
         """Get crypto balance."""
         pass
+
+
+class BrokerException(Exception):
+    """Base class for all broker-related exceptions."""
+    pass
+
+
+class OrderSubmissionError(BrokerException):
+    """Generic exception for order submission error."""
+    pass
+
+
+class InvalidOrderData(OrderSubmissionError):
+    """Order submission failed because order data is invalid."""
+    pass
+
+
+class InsufficientFunds(OrderSubmissionError):
+    """Order submission failed due to insufficient funds."""
+    pass
+
+
+class OrderCancellationError(BrokerException):
+    """Generic exception for order cancellation error."""
+    pass
 
 
 class AbstractBroker(ABC):
@@ -231,7 +320,7 @@ class AbstractBroker(ABC):
         pass
 
     @abstractmethod
-    def iter_trades(self) -> t.Iterator[Trade]:
+    def iter_trades(self) -> t.Iterator[TradeInfo]:
         """Get trades iterator."""
         pass
 
@@ -241,26 +330,19 @@ class AbstractBroker(ABC):
         pass
 
     @abstractmethod
-    def get_trades(self) -> t.Sequence[Trade]:
+    def get_trades(self) -> t.Sequence[TradeInfo]:
         """Get trades sequence."""
         pass
 
     @abstractmethod
-    def submit_order(self, order_factory: OrderFactory) -> OrderInfo:
-        """Submit order for execution."""
-        pass
-
-    @abstractmethod
-    def submit_market_order(
-                self, 
-                order_factory: MarketOrderFactory) -> OrderInfo:
+    def submit_market_order(self, 
+                            options: MarketOrderOptions) -> MarketOrderInfo:
         """Submit market order."""
         pass
 
     @abstractmethod
-    def submit_limit_order(
-                self, 
-                order_factory: LimitOrderFactory) -> LimitOrderInfo:
+    def submit_limit_order(self, 
+                           options: LimitOrderOptions) -> LimitOrderInfo:
         """Submit limit order."""
         pass
 
@@ -268,7 +350,7 @@ class AbstractBroker(ABC):
     def submit_take_profit_order(
                 self, 
                 order_side: OrderSide,
-                order_factory: TakeProfitFactory) -> StrategyOrderInfo:
+                options: TakeProfitOptions) -> TakeProfitInfo:
         """Submit Take Profit order."""
         pass
 
@@ -276,7 +358,7 @@ class AbstractBroker(ABC):
     def submit_stop_loss_order(
                 self, 
                 order_side: OrderSide,
-                order_factory: StopLossFactory) -> StrategyOrderInfo:
+                options: StopLossOptions) -> StopLossInfo:
         """Submit Stop Loss order."""
         pass
