+++ backintime/broker/default/__init__.py
@@ -1,0 +1,1 @@
+from .proxy import BrokerProxy
