+++ backintime/trading_strategy.py
@@ -6,13 +6,15 @@ from .candles import Candles
 from .timeframes import Timeframes
 from .analyser.analyser import Analyser
 from .analyser.indicators.base import IndicatorParam
-from .broker_proxy import BrokerProxy, OrderInfo, LimitOrderInfo
 from .broker.base import (
     OrderSide,
-    MarketOrderFactory,
-    LimitOrderFactory,
-    TakeProfitFactory,
-    StopLossFactory
+    MarketOrderOptions,
+    MarketOrderInfo,
+    LimitOrderOptions,
+    LimitOrderInfo,
+    TakeProfitOptions,
+    StopLossOptions,
+    AbstractBroker
 )
 
 
@@ -48,7 +50,7 @@ class TradingStrategy(ABC):
     candle_timeframes: t.Set[Timeframes] = set()
 
     def __init__(self, 
-                 broker: BrokerProxy,
+                 broker: AbstractBroker,
                  analyser: Analyser,
                  candles: Candles):
         self.broker=broker
@@ -63,44 +65,44 @@ class TradingStrategy(ABC):
     def position(self) -> Decimal:
         return self.broker.balance.available_crypto_balance
 
-    def buy(self, amount: t.Optional[Decimal] = None) -> OrderInfo:
+    def buy(self, amount: t.Optional[Decimal] = None) -> MarketOrderInfo:
         """Shortcut for submitting market buy order."""
         order_amount = amount or self.broker.max_fiat_for_taker
-        order = MarketOrderFactory(OrderSide.BUY, order_amount)
-        return self.broker.submit_market_order(order)
+        options = MarketOrderOptions(OrderSide.BUY, amount=order_amount)
+        return self.broker.submit_market_order(options)
 
-    def sell(self, amount: t.Optional[Decimal] = None) -> OrderInfo:
+    def sell(self, amount: t.Optional[Decimal] = None) -> MarketOrderInfo:
         """Shortcut for submitting market sell order."""
         order_amount = amount or self.position
-        order = MarketOrderFactory(OrderSide.SELL, order_amount)
-        return self.broker.submit_market_order(order)
+        options = MarketOrderOptions(OrderSide.SELL, amount=order_amount)
+        return self.broker.submit_market_order(options)
 
-    def limit_buy(self, 
+    def limit_buy(self,
                   order_price: Decimal,
-                  take_profit_factory: t.Optional[TakeProfitFactory] = None,
-                  stop_loss_factory: t.Optional[StopLossFactory] = None,
+                  take_profit: t.Optional[TakeProfitOptions] = None,
+                  stop_loss: t.Optional[StopLossOptions] = None,
                   amount: t.Optional[Decimal] = None) -> LimitOrderInfo:
         """Shortcut for submitting limit buy order."""
         order_amount = amount or self.broker.max_fiat_for_maker
-        order = LimitOrderFactory(OrderSide.BUY,
-                                  order_amount,
-                                  order_price,
-                                  take_profit_factory,
-                                  stop_loss_factory)
-        return self.broker.submit_limit_order(order)
+        options = LimitOrderOptions(OrderSide.BUY,
+                                    amount=order_amount,
+                                    order_price=order_price,
+                                    take_profit=take_profit,
+                                    stop_loss=stop_loss)
+        return self.broker.submit_limit_order(options)
 
     def limit_sell(self, 
                    order_price: Decimal,
-                   take_profit_factory: t.Optional[TakeProfitFactory] = None,
-                   stop_loss_factory: t.Optional[StopLossFactory] = None,
+                   take_profit: t.Optional[TakeProfitOptions] = None,
+                   stop_loss: t.Optional[StopLossOptions] = None,
                    amount: t.Optional[Decimal] = None) -> LimitOrderInfo:
         """Shortcut for submitting limit sell order."""
         order_amount = amount or self.position
-        order = LimitOrderFactory(OrderSide.SELL,
-                                  order_amount,
-                                  order_price,
-                                  take_profit_factory,
-                                  stop_loss_factory)
+        options = LimitOrderOptions(OrderSide.SELL,
+                                    amount=order_amount,
+                                    order_price=order_price,
+                                    take_profit=take_profit,
+                                    stop_loss=stop_loss)
         return self.broker.submit_limit_order(order)
 
     @abstractmethod
